articlescrape/requirements.txt
grpcio-tools
requests
readability-lxml
grpc-interceptor
selenium
breadability
python-dotenv
